Update best model creation file
This commit should really be applied to all model creation files, but I
was running out of time before MM started. This breaks out the training
of the model so I can train a model with seed knowledge and one without
as an interesting experiment with seed bias.

It also adds a fix to even out the TEAM1_WIN discrepancy. Previously,
the way the data was created the majority of TEAM1_WINs were 1, which
made the model naturally biased to pick 1 the heavy majority of the
time. This makes it even counts, so it actually uses the appropriate
features to make a decision.